add second time feature.
there are not other default uses ( like transit, public,
neighbourhood..) so, for 11pm, only have the flow between retail and
resident

+++ Main/UseFlow.pde
@@ -2,10 +2,10 @@ class UseFlow
 {
   String from_use;
   String to_use;
-  float time;
+  int time;
   float density;
   
-  UseFlow (float time_select, String from, String to, float number)
+  UseFlow (int time_select, String from, String to, float number)
   {
     time = time_select;
     from_use = from;
